Updated menuRequestHandler, responses and communication with client on highScore and personalStats works

+++ triviaProject/Project1/RequestInfo.h
@@ -1,9 +1,10 @@
 #pragma once
 #include <ctime>
 #include <vector>
+#include "Responses.h"
 
 struct RequestInfo {
-	unsigned int code;
+	BYTE code;
 	time_t receivalTime;
 	std::vector<unsigned char> buffer;
 };
